refactor: unify event and calendar mode event-title
